backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
python-multipart
google-generativeai
pdfplumber
langchain
langchain-text-splitters
langchain-google-genai
pydantic-settings
passlib
python-jose
python-dotenv
transformers
torch
sentence-transformers
SentencePiece
accelerate
python-docx
openpyxl
pandas
tenacity
pytesseract
asyncpg
psycopg2-binary
latex2mathml
lxml
